Initial Commit.
Initial Commit.

Some of this probably won't work as expected but the main part of refreshing every 30 minutes to 60 minutes and the logging to console works as expected. So core functionality works.

Miscellaneous features I'm unsure whether they work or not. If someone wants to test and or fix go for it.

As a side note. No clue whether requirements.txt works as expected. However even if it doesn't it should give you a list of what packages are needed.